moved normalize_timestamps to source_app/utils.py
